Move database.js from data to controller
